+++ tartufo/cli.py
@@ -45,7 +45,6 @@ class TartufoCLI(click.MultiCommand):
     name="tartufo",
     context_settings=dict(help_option_names=["-h", "--help"]),
 )
-@click.option("--json/--no-json", help="Output in JSON format.", is_flag=True)
 @click.option(
     "--rules",
     multiple=True,
@@ -59,13 +58,6 @@ class TartufoCLI(click.MultiCommand):
     show_default=True,
     help="Whether to include the default regex list when configuring"
     " search patterns. Only applicable if --rules is also specified.",
-)
-@click.option(
-    "--compact/--no-compact",
-    is_flag=True,
-    default=False,
-    show_default=True,
-    help="Enable reduced output.",
 )
 @click.option(
     "--entropy/--no-entropy",
@@ -121,6 +113,13 @@ class TartufoCLI(click.MultiCommand):
     exclude multiple patterns. If not provided (default), no Git object paths
     are excluded unless effectively excluded via the --include-path-patterns
     option.""",
+)
+@click.option(
+    "-of",
+    "--output-format",
+    type=click.Choice(["json", "compact", "text"]),
+    default="text",
+    help="Output can be in json/compact/text format. By default the output will be in text format.",
 )
 @click.option(
     "-xe",