+++ tartufo/types.py
@@ -7,7 +7,6 @@ from typing import Any, Dict, Optional, TextIO, Tuple, Pattern
 @dataclass
 class GlobalOptions:
     __slots__ = (
-        "json",
         "rules",
         "default_regexes",
         "entropy",
@@ -25,11 +24,10 @@ class GlobalOptions:
         "verbose",
         "quiet",
         "log_timestamps",
-        "compact",
+        "output_format",
         "b64_entropy_score",
         "hex_entropy_score",
     )
-    json: bool
     rules: Tuple[TextIO, ...]
     default_regexes: bool
     entropy: bool
@@ -47,7 +45,7 @@ class GlobalOptions:
     verbose: int
     quiet: bool
     log_timestamps: bool
-    compact: bool
+    output_format: Optional[str]
     b64_entropy_score: float
     hex_entropy_score: float