+++ tests/test_base_scanner.py
@@ -152,7 +152,7 @@ class IssueTests(unittest.TestCase):
             ),
         )
         issue.issue_detail = "issue-detail"
-        actual = issue.as_dict(compact=True)
+        actual = issue.as_dict(output_format="compact")
         self.assertEqual(
             {
                 "file_path": "test-file",